Permissões, adicionando legenda em campo
- Adicionando checagem de permissão em um form
- Modificações necessarias para caber legenda em campo
- Limpando o arquivo da view de seções movendo boa parte do js para um
  arquivo a parte
